requirements.txt
pytest~=6.2.5
pandas~=1.3.5
requests~=2.26.0
regex~=2021.11.10
Unidecode~=1.3.2
nltk~=3.6.5
tensorflow~=2.7.0
numpy~=1.21.4
sklearn~=0.0
scikit-learn~=1.0.1
python-dotenv~=0.19.2